renamed rewrite to just yellbot
